backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pymongo==4.6.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
httpx==0.25.2
aiofiles==23.2.1
PyPDF2==3.0.1
beautifulsoup4==4.12.2
lxml==4.9.3

